[Feat] DetailBookstoreView를 MVC 패턴으로 정리
